Create new useDefaultOutputDir() hook
